requirements.txt
PyYAML==3.12
pytest==3.0.7
pytest-cov==2.5.0
scipy==1.0.0
prettytable==0.7.2